Fix bug with not-overwritten folders during instantiating template
Since the tool 'cp' was called without the option '--no-target-directory',
there was a bug when instantiation was done onto a folder where
instantiation had been done earlier: the folder 'ote' was copied into
the subfolder 'packages/ote/ote' instead of the folder 'packages/ote'
itself.

This change fixes this issue.